Crate interfaces

A library for interoperating with the network interfaces of a system.

TODO: add more documentation on how to use.

Re-exports

pub use flags::InterfaceFlags;

Modules

flags
Submodule containing various flags.

Structs

Address
This structure represents a single address for a given interface.

HardwareAddr
HardwareAddr represents a hardware address (commonly known as a MAC address) of a given interface.

Interface
The Interface structure represents a single interface on the system. It also contains methods to control the interface.

Enums

InterfacesError
InterfacesError is the error type that is returned by all functions in this crate. See the documentation on the individual variants for more information.

Kind
Kind represents the interface family (equivalent to the sa_family field in the sockaddr structure).

NextHop
The next hop for an interface. See the individual variants for more information.

Type Definitions

Result
A specialized Result type for this crate.