Enum interfaces::InterfacesError

pub enum InterfacesError {
    Errno(Errno),
    NotSupported(&'static str),
}

InterfacesError is the error type that is returned by all functions in this crate. See the documentation on the individual variants for more information.

Variants
Errno(Errno)

Errno indicates that something went wrong with an underlying syscall. The internal value is the errno that was returned.

NotSupported(&'static str)

NotSupported indicates that something required for this operation is not currently supported on this platform or computer. The internal string may contain more detail.

Methods

impl InterfacesError

pub fn last_os_error() -> InterfacesError

Create a new instance of InterfacesError with the error set to the current value of the libc errno variable.