Module interfaces::flags

Submodule containing various flags.

Structs

InterfaceFlags
Represents a set of flags that describe the state of an interface. This corresponds to the flags that are returned from the SIOCGIFFLAGS syscall

Constants

IFF_ALLMULTI
Receive all multicast packets.

IFF_AUTOMEDIA
Auto media select active.

IFF_BROADCAST
Broadcast address valid.

IFF_DEBUG
Turn on debugging.

IFF_DYNAMIC
Dialup device with changing addresses.

IFF_LOOPBACK
Is a loopback net.

IFF_MASTER
Master of a load balancer.

IFF_MULTICAST
Supports multicast.

IFF_NOARP
No address resolution protocol.

IFF_NOTRAILERS
Avoid use of trailers.

IFF_POINTOPOINT
Interface is point-to-point link.

IFF_PORTSEL
Can set media type.

IFF_PROMISC
Receive all packets.

IFF_RUNNING
Resources allocated.

IFF_SLAVE
Slave of a load balancer.

IFF_UP
Interface is up.